Create task for preparing static data from easyway
