+++ way_to_home/way_to_home/celery.py
@@ -6,5 +6,5 @@ from celery import Celery
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'way_to_home.settings')
 
-CELERY_APP = Celery('WayToHome', include=['utils.notification_tasks'])
+CELERY_APP = Celery('WayToHome', include=['utils.tasks'])
 CELERY_APP.config_from_object('django.conf:settings', namespace='CELERY')
